backend_vNext/requirements.txt
# Backend vNext Dependencies
# Python 3.10+

# =============================================================================
# Web Framework
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sse-starlette>=1.8.0

# =============================================================================
# Database
# =============================================================================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# =============================================================================
# Task Queue (Background Jobs)
# =============================================================================
celery>=5.3.0
redis>=5.0.0

# =============================================================================
# LLM APIs
# =============================================================================
google-generativeai>=0.8.0
openai>=1.12.0
anthropic>=0.75.0

# =============================================================================
# Configuration & Validation
# =============================================================================
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1

# =============================================================================
# PDF Processing
# =============================================================================
PyMuPDF>=1.23.0
pillow>=10.0.0
agentic-doc

# =============================================================================
# PDF Annotation & Text Matching
# =============================================================================
rapidfuzz>=3.0.0
pytesseract>=0.3.10

# =============================================================================
# HTML/XML Parsing
# =============================================================================
beautifulsoup4>=4.12.0
lxml>=5.0.0

# =============================================================================
# JSON Schema Validation
# =============================================================================
jsonschema>=4.21.0

# =============================================================================
# HTTP Client & Utilities
# =============================================================================
httpx>=0.26.0
requests>=2.31.0
aiofiles>=23.2.0
tenacity>=8.2.0
tqdm>=4.66.0

# =============================================================================
# YAML Configuration
# =============================================================================
PyYAML>=6.0.0

# =============================================================================
# Authentication (JWT)
# =============================================================================
PyJWT>=2.8.0

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
pydantic[email]